requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
environs==9.5.0
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi_cors==0.0.6
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
nest-asyncio==1.5.7
packaging==23.1
pydantic==2.3.0
pydantic_core==2.6.3
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
